Important bug fix with schema change
Re-generating a user token (e.g. for an existing account clicking
"send me an auth email" again) would create a new account instead of
sending a uuid that references the existing account.

To fix, we need to look up a user uuid given the email address. So we
create a new index mapping emails back to user uuid.