Large refactor of email notifier
First in a series of steps to upgrade the notifier. This change was primarily a
refactor with no new externally visible functionality. However, the interals
were made more modular so that the source of sonde data can be mocked out,
allowing the email notifier to work on replayed data.

First notifier test written that verifies an actual email is sent using moto to
mock the AWS email sending service (SES).