Make notifier more resilient to sondehub data errors
The notifier began to fail because 'vel_v' could not be converted to a
float. While the exception was recorded, I unfortunately wasn't able
to capture what value actually caused the failure.

This PR changes the notifier to handle the situation more gracefully,
and print the offending dataframe so I can perhaps see what the bad
data was.